requirements.txt
Django==3.1
Pillow==8.1
psycopg2-binary==2.8.6
djangorestframework==3.11.1
python-decouple
django-ckeditor
pytils
drf-yasg
django-filter
